amazonka: Split instance profile auth into its own module
