+++ lib/amazonka/BUILD.bazel
@@ -7,7 +7,9 @@ haskell_library(
         "src/Amazonka.hs",
         "src/Amazonka/Auth.hs",
         "src/Amazonka/Auth.hs-boot",  # keep
+        "src/Amazonka/Auth/Background.hs",
         "src/Amazonka/Auth/Exception.hs",
+        "src/Amazonka/Auth/InstanceProfile.hs",
         "src/Amazonka/Auth/Keys.hs",
         "src/Amazonka/EC2/Metadata.hs",
         "src/Amazonka/Env.hs",
