+++ lib/amazonka/src/Amazonka/Auth.hs
@@ -71,7 +71,9 @@ module Amazonka.Auth
   )
 where
 
+import Amazonka.Auth.Background (fetchAuthInBackground)
 import Amazonka.Auth.Exception
+import Amazonka.Auth.InstanceProfile (fromProfile, fromProfileName)
 import Amazonka.Auth.Keys (fromKeys, fromKeysEnv, fromSession, fromTemporarySession)
 import Amazonka.Data
 import Amazonka.EC2.Metadata
@@ -81,31 +83,23 @@ import Amazonka.Prelude
 import qualified Amazonka.STS as STS
 import qualified Amazonka.STS.AssumeRoleWithWebIdentity as STS
 import Amazonka.Types
-import Control.Concurrent (ThreadId)
-import qualified Control.Concurrent as Concurrent
 import qualified Control.Exception as Exception
 import Control.Lens ((^.))
-import Control.Monad.Catch (MonadThrow(..))
+import Control.Monad.Catch (MonadThrow (..))
 import Control.Monad.Trans.Maybe (MaybeT (..))
 import Control.Monad.Trans.Resource (runResourceT)
 import qualified Data.ByteString.Char8 as BS8
-import qualified Data.ByteString.Lazy.Char8 as LBS8
 import qualified Data.Char as Char
-import Data.IORef (IORef)
-import qualified Data.IORef as IORef
 import qualified Data.Ini as INI
 import Data.Monoid (Dual, Endo)
 import qualified Data.Text as Text
 import qualified Data.Text.Encoding as Text
 import qualified Data.Text.IO as Text
-import qualified Data.Time as Time
 import qualified Data.UUID as UUID
 import qualified Data.UUID.V4 as UUID
 import qualified Network.HTTP.Client as Client
 import qualified System.Directory as Directory
 import qualified System.Environment as Environment
-import System.Mem.Weak (Weak)
-import qualified System.Mem.Weak as Weak
 
 -- | Default access key environment variable.
 envAccessKey ::
@@ -309,8 +303,9 @@ instance Show Credentials where
 runCredentialChain :: MonadThrow m => [(a -> m (Maybe b))] -> a -> m b
 runCredentialChain chain env = case chain of
   [] -> throwM CredentialChainExhausted
-  provider : chain' -> provider env
-    >>= maybe (runCredentialChain chain' env) pure
+  provider : chain' ->
+    provider env
+      >>= maybe (runCredentialChain chain' env) pure
 
 -- | Retrieve authentication information via the specified 'Credentials' mechanism.
 --
@@ -432,79 +427,6 @@ fromFilePath profile cred conf =
           <> maybe mempty (", key " <>) mkey
           <> ", "
           <> Text.pack err
-
--- | Retrieve the default IAM Profile from the local EC2 instance-data.
---
--- The default IAM profile is determined by Amazon as the first profile found
--- in the response from:
--- @http://169.254.169.254/latest/meta-data/iam/security-credentials/@
---
--- Throws 'RetrievalError' if the HTTP call fails, or 'InvalidIAMError' if
--- the default IAM profile cannot be read.
-fromProfile ::
-  MonadIO m =>
-  Client.Manager ->
-  m (Auth, Maybe Region)
-fromProfile m =
-  liftIO $ do
-    ls <- Exception.try $ metadata m (IAM (SecurityCredentials Nothing))
-
-    case BS8.lines <$> ls of
-      Right (x : _) -> fromProfileName m (Text.decodeUtf8 x)
-      Left e -> Exception.throwIO (RetrievalError e)
-      _ ->
-        Exception.throwIO $
-          InvalidIAMError "Unable to get default IAM Profile from EC2 metadata"
-
--- | Lookup a specific IAM Profile by name from the local EC2 instance-data.
---
--- Additionally starts a refresh thread for the given authentication environment.
---
--- The resulting 'IORef' wrapper + timer is designed so that multiple concurrent
--- accesses of 'AuthEnv' from the 'AWS' environment are not required to calculate
--- expiry and sequentially queue to update it.
---
--- The forked timer ensures a singular owner and pre-emptive refresh of the
--- temporary session credentials before expiration.
---
--- A weak reference is used to ensure that the forked thread will eventually
--- terminate when 'Auth' is no longer referenced.
---
--- If no session token or expiration time is present the credentials will
--- be returned verbatim.
-fromProfileName ::
-  MonadIO m =>
-  Client.Manager ->
-  Text ->
-  m (Auth, Maybe Region)
-fromProfileName m name =
-  liftIO $ do
-    auth <- fetchAuthInBackground getCredentials
-    reg <- getRegionFromIdentity
-
-    pure (auth, Just reg)
-  where
-    getCredentials =
-      Exception.try (metadata m (IAM . SecurityCredentials $ Just name))
-        >>= handleErr (eitherDecode' . LBS8.fromStrict) invalidIAMErr
-
-    getRegionFromIdentity =
-      Exception.try (identity m)
-        >>= handleErr (fmap _region) invalidIdentityErr
-
-    handleErr f g = \case
-      Left e -> Exception.throwIO (RetrievalError e)
-      Right x -> either (Exception.throwIO . g) pure (f x)
-
-    invalidIAMErr =
-      InvalidIAMError
-        . mappend ("Error parsing IAM profile '" <> name <> "' ")
-        . Text.pack
-
-    invalidIdentityErr =
-      InvalidIAMError
-        . mappend "Error parsing Instance Identity Document "
-        . Text.pack
 
 -- | https://aws.amazon.com/blogs/opensource/introducing-fine-grained-iam-roles-service-accounts/
 -- Obtain temporary credentials from STS:AssumeRoleWithWebIdentity
@@ -643,48 +565,6 @@ getRegion = runMaybeT $ do
     (const . MaybeT $ pure Nothing)
     pure
     (fromText (Text.pack mr))
-
--- | Implements the background fetching behavior used by 'fromProfileName' and
--- 'fromContainer'. Given an 'IO' action that produces an 'AuthEnv', this spawns
--- a thread that mutates the 'IORef' returned in the resulting 'Auth' to keep
--- the temporary credentials up to date.
-fetchAuthInBackground :: IO AuthEnv -> IO Auth
-fetchAuthInBackground menv =
-  menv >>= \(!env) -> liftIO $
-    case _authExpiration env of
-      Nothing -> pure (Auth env)
-      Just x -> do
-        r <- IORef.newIORef env
-        p <- Concurrent.myThreadId
-        s <- timer menv r p x
-
-        pure (Ref s r)
-  where
-    timer :: IO AuthEnv -> IORef AuthEnv -> ThreadId -> ISO8601 -> IO ThreadId
-    timer ma !r !p !x =
-      Concurrent.forkIO $ do
-        s <- Concurrent.myThreadId
-        w <- IORef.mkWeakIORef r (Concurrent.killThread s)
-
-        loop ma w p x
-
-    loop :: IO AuthEnv -> Weak (IORef AuthEnv) -> ThreadId -> ISO8601 -> IO ()
-    loop ma w !p !x = do
-      diff x <$> Time.getCurrentTime >>= Concurrent.threadDelay
-      env <- Exception.try ma
-      case env of
-        Left e -> Exception.throwTo p (RetrievalError e)
-        Right !a -> do
-          mr <- Weak.deRefWeak w
-          case mr of
-            Nothing -> pure ()
-            Just r -> do
-              IORef.atomicWriteIORef r a
-              maybe (pure ()) (loop ma w p) (_authExpiration a)
-
-    diff (Time !x) !y = (* 1000000) $ if n > 0 then n else 1
-      where
-        !n = truncate (Time.diffUTCTime x y) - 60
 
 -- $env
 -- This really should be defined in @Amazonka.Env@, but we define it
